requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pymysql>=1.0.2
PyYAML>=6.0
requests>=2.28.1
python-multipart>=0.0.5
pydub>=0.25.1
PyJWT>=2.8.0
email-validator>=2.0.0
bcrypt>=4.0.0
